Move entry details header to separate widget
