+++ lotti/pubspec.yaml
@@ -1,7 +1,7 @@
 name: lotti
 description: A Smart Journal.
 publish_to: 'none'
-version: 0.3.45+265
+version: 0.3.46+266
 
 environment:
   sdk: ">=2.14.0 <3.0.0"
